requirement.txt
pandas
psycopg2-binary
python-dotenv
requests
pdfminer.six
